+++ examples/texture.rs
@@ -7,10 +7,6 @@ struct GameState {
 }
 
 impl State for GameState {
-    fn update(&mut self, _ctx: &mut Context) -> tetra::Result {
-        Ok(())
-    }
-
     fn draw(&mut self, ctx: &mut Context, _dt: f64) -> tetra::Result {
         graphics::clear(ctx, Color::rgb(0.769, 0.812, 0.631));
         graphics::draw(ctx, &self.texture, Vec2::new(16.0, 16.0));