+++ examples/texture.rs
@@ -5,6 +5,14 @@ struct GameState {
     texture: Texture,
 }
 
+impl GameState {
+    fn new(ctx: &mut Context) -> tetra::Result<GameState> {
+        Ok(GameState {
+            texture: Texture::new(ctx, "./examples/resources/player.png")?,
+        })
+    }
+}
+
 impl State for GameState {
     fn draw(&mut self, ctx: &mut Context, _dt: f64) -> tetra::Result {
         graphics::clear(ctx, Color::rgb(0.769, 0.812, 0.631));
@@ -15,15 +23,10 @@ impl State for GameState {
 }
 
 fn main() -> tetra::Result {
-    let ctx = &mut ContextBuilder::new("Rendering a Texture", 160, 144)
+    ContextBuilder::new("Rendering a Texture", 160, 144)
         .maximized(true)
         .resizable(true)
         .quit_on_escape(true)
-        .build()?;
-
-    let state = &mut GameState {
-        texture: Texture::new(ctx, "./examples/resources/player.png")?,
-    };
-
-    tetra::run(ctx, state)
+        .build()?
+        .run_with(GameState::new)
 }