+++ CHANGELOG.md
@@ -6,12 +6,13 @@
 
 * The library has been upgraded to the 2018 edition of Rust.
 * `ContextBuilder::new` now takes the title and size as parameters. The old behavior of the function can be replicated by using `ContextBuilder::default` instead.
+* `run` is now a method on `Context`, instead of a free function. 
 * The `update` and `draw` methods on `State` now return `tetra::Result`, allowing errors to be returned (or propagated via the `?` operator). Any errors returned from these methods will stop the game - your main method can then handle the error (e.g. log it out). ([#29](https://github.com/17cupsofcoffee/tetra/issues/29))
 * The `scale` option on `ContextBuilder` has been renamed to `window_scale`, to better reflect its behavior.
 * `Shader::from_file` is now called `Shader::new`, and `Shader::new` is now called `Shader::from_string`. This is more consistent with the other constructors.
 * Tick rates are now specified in ticks per second ([#40](https://github.com/17cupsofcoffee/tetra/issues/40)).
 * The `ContextBuilder` no longer consumes itself when called - this is more flexible for e.g. calling methods inside a conditional.
-* `set_tick_rate` has been moved to the `window` module.
+* `quit` and `set_tick_rate` have been moved to the `window` module.
 * The functions for getting the game's internal width/height have been renamed to disambiguate them from the functions for getting the window width/height.
 
 ### New Features
@@ -30,6 +31,7 @@
 * There are now many new methods for manipulating the window/game loop in the `window` module.
 * The `update` and `draw` methods on `State` are now both optional.
 * The `graphics` module now re-exports `Vec2`.
+* In addition to the normal `run` method, there is now also a `run_with` method that uses a closure to construct the `State`. This is handy when method chaining - see the examples for how it can be used.
 
 ### Bug Fixes