+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * The `ContextBuilder` no longer consumes itself when called - this is more flexible for e.g. calling methods inside a conditional.
 * `quit` and `set_tick_rate` have been moved to the `window` module.
 * The functions for getting the game's internal width/height have been renamed to disambiguate them from the functions for getting the window width/height.
+* Matching on `TetraError` will now force you to add a wildcard arm. This will prevent the addition of new error types from being a breaking change.
 
 ### New Features
 
@@ -33,6 +34,7 @@
 * The `graphics` module now re-exports `Vec2`.
 * In addition to the normal `run` method, there is now also a `run_with` method that uses a closure to construct the `State`. This is handy when method chaining - see the examples for how it can be used.
 * Public types now implement `Debug` and `Clone` where appropriate.
+* `TetraError` now implements the standard library `Error` trait ([#46](https://github.com/17cupsofcoffee/tetra/issues/46)).
 
 ### Bug Fixes